requirements.txt
locustio==0.8.1
bs4==0.0.1